corrected comments, remove getTemplate function, added renderPictures function
